findbugs: fixing boxing-unboxing issues

+++ frontend/webadmin/modules/uicommonweb/src/main/java/org/ovirt/engine/ui/uicommonweb/dataprovider/AsyncDataProvider.java
@@ -775,7 +775,7 @@ public final class AsyncDataProvider {
                 {
                     return 100;
                 }
-                return ((Integer) source).intValue();
+                return source;
             }
         };
         GetConfigFromCache(
